+++ hostdiscovery/hostdiscovery_test.go
@@ -1,3 +1,5 @@
+// +build libnetwork_discovery
+
 package hostdiscovery
 
 import (
